Date:   Thu Jan 22 19:41:41 2015 +0100
    Few double compatibility changes.

Date:   Sun Jan 18 23:59:29 2015 +0100
    Nearly ended, but still a trouble with a little deformation of the mesh...

Date:   Fri Jan 16 17:49:20 2015 +0100
    Something weird for the new point position, but the algorithm is there.

+++ source/Core/CastorUtils/Src/Font.cpp
@@ -177,7 +177,7 @@ namespace Castor
 					uint8_t * l_dst = l_bitmap.data();
 					uint8_t const * l_src = l_ftBitmap.buffer;
 
-					for ( unsigned int i = 0; i < l_ftBitmap.rows; i++ )
+					for ( uint32_t i = 0; i < uint32_t( l_ftBitmap.rows ); i++ )
 					{
 						memcpy( l_dst, l_src, l_ftBitmap.width );
 						l_src += l_pitch;
@@ -188,8 +188,8 @@ namespace Castor
 				{
 					uint8_t * l_dst = l_bitmap.data() + l_uiSize - l_pitch;
 					uint8_t const * l_src = l_ftBitmap.buffer;
-
-					for ( unsigned int i = 0; i < l_ftBitmap.rows; i++ )
+					
+					for ( uint32_t i = 0; i < uint32_t( l_ftBitmap.rows ); i++ )
 					{
 						memcpy( l_dst, l_src, l_ftBitmap.width );
 						l_src += l_pitch;
